Allow engineers to easily mix and match the various existing filter methods to create filters however they wish, e.g. a standard file type with custom fields
