+++ src/Browscap/Filter/Section/LiteSectionFilter.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Browscap\Filter\Section;
+
+/**
+ * this filter is responsible to select properties and sections for the "full" version of the browscap files
+ */
+trait LiteSectionFilter
+{
+    /**
+     * checks if a section should be in the output
+     *
+     * @param bool[] $section
+     *
+     * @throws void
+     */
+    public function isOutputSection(array $section): bool
+    {
+        return isset($section['lite']) && $section['lite'];
+    }
+}
